\name{computeYield}
\alias{computeYield}
\title{This function computes the yield given area and production}
\usage{
  computeYield(production, area)
}
\arguments{
  \item{production}{The production time series}

  \item{area}{The area time series}
}
\description{
  This function avoids Inf, NAN when calculating the yield
}

