\name{splitNACountry}
\alias{splitNACountry}
\title{This function splits the data into two data.frame depending on
whether the country contain any data.}
\usage{
  splitNACountry(value, country, data)
}
\arguments{
  \item{value}{The column name for the variable of
  interest.}

  \item{country}{The column name of the country.}

  \item{data}{the name of the data}
}
\description{
  This function splits the data into two data.frame
  depending on whether the country contain any data.
}

